+++ types/decentlab/PR21/uplink.spec.js
@@ -26,7 +26,7 @@ before(function (done) {
       if (err) throw err;
       defaultSchema = JSON.parse(fileContents);
       done();
-    }
+    },
   );
 });
 
@@ -38,7 +38,7 @@ before(function (done) {
       if (err) throw err;
       lifecycleSchema = JSON.parse(fileContents);
       done();
-    }
+    },
   );
 });
 
@@ -56,7 +56,7 @@ describe("Decentlab PR21 Uplink", function () {
     it("should decode Decentlab PR21 payload", function (done) {
       const data = {
         data: {
-          payload_hex: "02016700034e8060170c7f",
+          payloadHex: "02016700034e8060170c7f",
         },
       };