+++ types/decentlab/SMTP/uplink.js
@@ -3,104 +3,142 @@ var decentlab_decoder = {
   SENSORS: [
     {
       length: 16,
-      values: [{
-        name: 'soil_moisture_at_depth_0',
-        displayName: 'Soil moisture at depth 0',
-        convert: function (x) { return (x[0] - 2500) / 500; }
-      },
-      {
-        name: 'soil_temperature_at_depth_0',
-        displayName: 'Soil temperature at depth 0',
-        convert: function (x) { return (x[1] - 32768) / 100; },
-        unit: '°C'
-      },
-      {
-        name: 'soil_moisture_at_depth_1',
-        displayName: 'Soil moisture at depth 1',
-        convert: function (x) { return (x[2] - 2500) / 500; }
-      },
-      {
-        name: 'soil_temperature_at_depth_1',
-        displayName: 'Soil temperature at depth 1',
-        convert: function (x) { return (x[3] - 32768) / 100; },
-        unit: '°C'
-      },
-      {
-        name: 'soil_moisture_at_depth_2',
-        displayName: 'Soil moisture at depth 2',
-        convert: function (x) { return (x[4] - 2500) / 500; }
-      },
-      {
-        name: 'soil_temperature_at_depth_2',
-        displayName: 'Soil temperature at depth 2',
-        convert: function (x) { return (x[5] - 32768) / 100; },
-        unit: '°C'
-      },
-      {
-        name: 'soil_moisture_at_depth_3',
-        displayName: 'Soil moisture at depth 3',
-        convert: function (x) { return (x[6] - 2500) / 500; }
-      },
-      {
-        name: 'soil_temperature_at_depth_3',
-        displayName: 'Soil temperature at depth 3',
-        convert: function (x) { return (x[7] - 32768) / 100; },
-        unit: '°C'
-      },
-      {
-        name: 'soil_moisture_at_depth_4',
-        displayName: 'Soil moisture at depth 4',
-        convert: function (x) { return (x[8] - 2500) / 500; }
-      },
-      {
-        name: 'soil_temperature_at_depth_4',
-        displayName: 'Soil temperature at depth 4',
-        convert: function (x) { return (x[9] - 32768) / 100; },
-        unit: '°C'
-      },
-      {
-        name: 'soil_moisture_at_depth_5',
-        displayName: 'Soil moisture at depth 5',
-        convert: function (x) { return (x[10] - 2500) / 500; }
-      },
-      {
-        name: 'soil_temperature_at_depth_5',
-        displayName: 'Soil temperature at depth 5',
-        convert: function (x) { return (x[11] - 32768) / 100; },
-        unit: '°C'
-      },
-      {
-        name: 'soil_moisture_at_depth_6',
-        displayName: 'Soil moisture at depth 6',
-        convert: function (x) { return (x[12] - 2500) / 500; }
-      },
-      {
-        name: 'soil_temperature_at_depth_6',
-        displayName: 'Soil temperature at depth 6',
-        convert: function (x) { return (x[13] - 32768) / 100; },
-        unit: '°C'
-      },
-      {
-        name: 'soil_moisture_at_depth_7',
-        displayName: 'Soil moisture at depth 7',
-        convert: function (x) { return (x[14] - 2500) / 500; }
-      },
-      {
-        name: 'soil_temperature_at_depth_7',
-        displayName: 'Soil temperature at depth 7',
-        convert: function (x) { return (x[15] - 32768) / 100; },
-        unit: '°C'
-      }]
+      values: [
+        {
+          name: "soil_moisture_at_depth_0",
+          displayName: "Soil moisture at depth 0",
+          convert: function (x) {
+            return (x[0] - 2500) / 500;
+          },
+        },
+        {
+          name: "soil_temperature_at_depth_0",
+          displayName: "Soil temperature at depth 0",
+          convert: function (x) {
+            return (x[1] - 32768) / 100;
+          },
+          unit: "°C",
+        },
+        {
+          name: "soil_moisture_at_depth_1",
+          displayName: "Soil moisture at depth 1",
+          convert: function (x) {
+            return (x[2] - 2500) / 500;
+          },
+        },
+        {
+          name: "soil_temperature_at_depth_1",
+          displayName: "Soil temperature at depth 1",
+          convert: function (x) {
+            return (x[3] - 32768) / 100;
+          },
+          unit: "°C",
+        },
+        {
+          name: "soil_moisture_at_depth_2",
+          displayName: "Soil moisture at depth 2",
+          convert: function (x) {
+            return (x[4] - 2500) / 500;
+          },
+        },
+        {
+          name: "soil_temperature_at_depth_2",
+          displayName: "Soil temperature at depth 2",
+          convert: function (x) {
+            return (x[5] - 32768) / 100;
+          },
+          unit: "°C",
+        },
+        {
+          name: "soil_moisture_at_depth_3",
+          displayName: "Soil moisture at depth 3",
+          convert: function (x) {
+            return (x[6] - 2500) / 500;
+          },
+        },
+        {
+          name: "soil_temperature_at_depth_3",
+          displayName: "Soil temperature at depth 3",
+          convert: function (x) {
+            return (x[7] - 32768) / 100;
+          },
+          unit: "°C",
+        },
+        {
+          name: "soil_moisture_at_depth_4",
+          displayName: "Soil moisture at depth 4",
+          convert: function (x) {
+            return (x[8] - 2500) / 500;
+          },
+        },
+        {
+          name: "soil_temperature_at_depth_4",
+          displayName: "Soil temperature at depth 4",
+          convert: function (x) {
+            return (x[9] - 32768) / 100;
+          },
+          unit: "°C",
+        },
+        {
+          name: "soil_moisture_at_depth_5",
+          displayName: "Soil moisture at depth 5",
+          convert: function (x) {
+            return (x[10] - 2500) / 500;
+          },
+        },
+        {
+          name: "soil_temperature_at_depth_5",
+          displayName: "Soil temperature at depth 5",
+          convert: function (x) {
+            return (x[11] - 32768) / 100;
+          },
+          unit: "°C",
+        },
+        {
+          name: "soil_moisture_at_depth_6",
+          displayName: "Soil moisture at depth 6",
+          convert: function (x) {
+            return (x[12] - 2500) / 500;
+          },
+        },
+        {
+          name: "soil_temperature_at_depth_6",
+          displayName: "Soil temperature at depth 6",
+          convert: function (x) {
+            return (x[13] - 32768) / 100;
+          },
+          unit: "°C",
+        },
+        {
+          name: "soil_moisture_at_depth_7",
+          displayName: "Soil moisture at depth 7",
+          convert: function (x) {
+            return (x[14] - 2500) / 500;
+          },
+        },
+        {
+          name: "soil_temperature_at_depth_7",
+          displayName: "Soil temperature at depth 7",
+          convert: function (x) {
+            return (x[15] - 32768) / 100;
+          },
+          unit: "°C",
+        },
+      ],
     },
     {
       length: 1,
-      values: [{
-        name: 'battery_voltage',
-        displayName: 'Battery voltage',
-        convert: function (x) { return x[0] / 1000; },
-        unit: 'V'
-      }]
-    }
+      values: [
+        {
+          name: "battery_voltage",
+          displayName: "Battery voltage",
+          convert: function (x) {
+            return x[0] / 1000;
+          },
+          unit: "V",
+        },
+      ],
+    },
   ],
 
   read_int: function (bytes, pos) {
@@ -110,7 +148,7 @@ var decentlab_decoder = {
   decode: function (msg) {
     var bytes = msg;
     var i, j;
-    if (typeof msg === 'string') {
+    if (typeof msg === "string") {
       bytes = [];
       for (i = 0; i < msg.length; i += 2) {
         bytes.push(parseInt(msg.substring(i, i + 2), 16));
@@ -124,12 +162,11 @@ var decentlab_decoder = {
 
     var deviceId = this.read_int(bytes, 1);
     var flags = this.read_int(bytes, 3);
-    var result = { 'protocol_version': version, 'device_id': deviceId };
+    var result = { protocol_version: version, device_id: deviceId };
     // decode payload
     var pos = 5;
     for (i = 0; i < this.SENSORS.length; i++, flags >>= 1) {
-      if ((flags & 1) !== 1)
-        continue;
+      if ((flags & 1) !== 1) continue;
 
       var sensor = this.SENSORS[i];
       var x = [];
@@ -142,18 +179,18 @@ var decentlab_decoder = {
       // decode sensor values
       for (j = 0; j < sensor.values.length; j++) {
         var value = sensor.values[j];
-        if ('convert' in value) {
+        if ("convert" in value) {
           result[value.name] = value.convert.bind(this)(x);
         }
       }
     }
     return result;
-  }
+  },
 };
 
 function deleteUnusedKeys(data) {
   var keysRetained = false;
-  Object.keys(data).forEach(key => {
+  Object.keys(data).forEach((key) => {
     if (data[key] === undefined) {
       delete data[key];
     } else {
@@ -164,7 +201,7 @@ function deleteUnusedKeys(data) {
 }
 
 function consume(event) {
-  var payload = event.data.payload_hex;
+  var payload = event.data.payloadHex;
   var sample = decentlab_decoder.decode(payload);
   var data = {};
   var lifecycle = {};
@@ -193,10 +230,10 @@ function consume(event) {
   lifecycle.deviceID = sample["device_id"];
 
   if (deleteUnusedKeys(data)) {
-    emit('sample', { "data": data, "topic": "default" });
+    emit("sample", { data: data, topic: "default" });
   }
 
   if (deleteUnusedKeys(lifecycle)) {
-    emit('sample', { "data": lifecycle, "topic": "lifecycle" });
+    emit("sample", { data: lifecycle, topic: "lifecycle" });
   }
 }