+++ types/talkpool/oy1210/uplink.js
@@ -1,9 +1,9 @@
 function consume(event) {
-  var payload = event.data.payload_hex;
-  var port = event.data.port;
-  var bits = Bits.hexToBits(payload);
-  var data = {};
-  var topic = "default";
+  const payload = event.data.payloadHex;
+  const { port } = event.data;
+  const bits = Bits.hexToBits(payload);
+  const data = {};
+  let topic = "default";
 
   if (port == 1) {
     if (payload == "012000") {
@@ -14,9 +14,19 @@ function consume(event) {
     topic = "startup";
   } else if (port == 2) {
     // Taking the temperature from different positions in the payload and converting it through the defined offsets (same for humidity)
-    data.temperature = Math.round((Bits.bitsToUnsigned(bits.substr(0, 8) + bits.substr(16, 4)) / 10 - 80) * 10) / 10;
-    data.humidity = Math.round((Bits.bitsToUnsigned(bits.substr(8, 8) + bits.substr(20, 4)) / 10 - 25) * 10) / 10;
+    data.temperature =
+      Math.round(
+        (Bits.bitsToUnsigned(bits.substr(0, 8) + bits.substr(16, 4)) / 10 -
+          80) *
+          10,
+      ) / 10;
+    data.humidity =
+      Math.round(
+        (Bits.bitsToUnsigned(bits.substr(8, 8) + bits.substr(20, 4)) / 10 -
+          25) *
+          10,
+      ) / 10;
     data.co2 = Bits.bitsToUnsigned(bits.substr(24, 16));
   }
-  emit('sample', { data: data, topic: topic });
+  emit("sample", { data, topic });
 }