+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+## [1.0.12] - 2021-07-20
+
+### Fixed
+
+- Comtac - CM 1 uplink
+
+## [1.0.11] - 2021-07-12
+
+### Added
+
+- Gavazzi - UWPA/UWPM
+
+## [1.0.10] - 2021-07-06
+
+### Changed
+
+- Deprecates the usage of `payload_hex` => use `payloadHex` instead
+
 ## [1.0.9] - 2021-06-22
 
 ### Added